% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_sql_queries.R
\name{combine_sql_queries}
\alias{combine_sql_queries}
\title{Combine SQL Queries}
\usage{
combine_sql_queries(
  arguments,
  argument_sequence,
  path_list,
  requested_vars = NULL
)
}
\arguments{
\item{arguments}{A list of SQL query arguments.}

\item{argument_sequence}{A numeric vector indicating the order of query arguments. Each number corresponds
to the index of the query argument in the \code{arguments} list.}

\item{path_list}{A list representing the join path. Each element of the list should be a data frame
describing a step in the join path with columns: "table_to_join", "method", and "common_var".}

\item{requested_vars}{A character vector specifying the variables to be selected from the final query result.
If \code{NULL}, all variables are selected.}
}
\value{
A combined SQL query string that integrates the specified queries using logical OR and AND operations.
}
\description{
This function combines multiple SQL queries into a single query using logical OR and AND operations.
It takes a list of SQL query arguments, an argument sequence indicating the order of arguments,
and a path list for query conversion. The queries are combined using OR operations within each group and
AND operations between different groups of queries.
}
