% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_fastest_way_to_table.R
\name{compute_fastest_way_to_table}
\alias{compute_fastest_way_to_table}
\title{Compute Fastest Way to Table}
\usage{
compute_fastest_way_to_table(
  conn,
  input_table = NULL,
  target_table = "observation_table"
)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{input_table}{The input table(s) from which the traversal begins. If \code{NULL}, all tables are considered as inputs.}

\item{target_table}{The target table that you want to reach through joins.}
}
\value{
A list of join paths with waypoint tables and walk approaches.
}
\description{
This function computes the fastest way to reach a target table from given input table(s) in a database using a combination
of forward and backward joins. It returns a list of join paths along with waypoint tables and walk approaches (forward or backward)
that enable efficient traversal from the input table(s) to the target table.
}
