% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continue_after_warning.R
\name{continue_after_warning}
\alias{continue_after_warning}
\title{Require user input after warning}
\usage{
continue_after_warning(message)
}
\arguments{
\item{message}{The message displayed to the user.}
}
\value{
Nothing. Throws an error depending on user input
}
\description{
This function requires the user to input that they want to continue with the current process.
}
