% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_query_path_to_sql.R
\name{convert_query_path_to_sql}
\alias{convert_query_path_to_sql}
\title{Convert Query Path to SQL}
\usage{
convert_query_path_to_sql(argument, path_list, requested_vars = NULL)
}
\arguments{
\item{argument}{The table-specific filtering argument.}

\item{path_list}{A list representing the join path. Each element of the list should be a data frame
describing a step in the join path with columns: "waypoint_tables", "walk_approaches", and "target_table".}

\item{requested_vars}{A character vector specifying the variables to be selected from the final query result.
If \code{NULL}, all variables are selected.}
}
\value{
A SQL query string that represents the desired filtering and joins on the target table.
}
\description{
This function converts a table-specific filtering argument specified in 'argument'
to a more general SQL query that returns entries from the target table that was
used to generate the path list. It uses the query to determine at which table the argument
is set and then constructs a SQL query by utilizing the table pathway to the target table.
}
