% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_db.R
\name{download_db}
\alias{download_db}
\title{Download the database file}
\usage{
download_db(local_path)
}
\arguments{
\item{local_path}{The local path where the downloaded file will be saved.}
}
\value{
NULL (no explicit return value).
}
\description{
This function downloads the database file from a specified URL and saves it to the local path.
}
\examples{
# Download the database file to the current working directory
download_db("initial_db.db")

}
