% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview_information.R
\name{get_overview_information}
\alias{get_overview_information}
\title{Get Overview Dataset Information}
\usage{
get_overview_information(conn, arguments, argument_relation)
}
\arguments{
\item{conn}{A database connection object.}

\item{arguments}{A list of arguments to filter the dataset.}

\item{argument_relation}{A logical operator ('AND' or 'OR') for combining multiple filter conditions.}
}
\value{
A data frame containing summary information for each dataset, indicating the presence of within
and between manipulations.
}
\description{
This function retrieves an overview of dataset information from a database connection.
}
