% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db_deprecated.R
\name{query_db_deprecated}
\alias{query_db_deprecated}
\title{Query the database based on specified arguments}
\usage{
query_db_deprecated(conn, arguments, argument_relation = "and")
}
\arguments{
\item{conn}{A database connection object.}

\item{arguments}{A list of SQL queries specifying the arguments.
Each query should retrieve the primary keys for matching records.}

\item{argument_relation}{The relation between the arguments. Defaults to "and".
Possible values include "and" and "or" and a custom vector.}
}
\value{
The requested data from the database based on the specified arguments.
}
\description{
This function queries the database based on the provided arguments and returns the requested data.
It matches primary keys from the initial queries and finds connected keys based on the argument sequence.
The data is then retrieved from the database using the obtained IDs.
}
\examples{
# Query the database
conn <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
arguments <- list(
  "SELECT observation_id FROM observation_table WHERE value > 10",
  "SELECT between_id FROM between_table WHERE between_var = 'A'"
)
query_db(conn, arguments)

}
