% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_db.R
\name{update_db}
\alias{update_db}
\title{Update the database file}
\usage{
update_db(local_path)
}
\arguments{
\item{local_path}{The local path where the downloaded file will be saved.}
}
\value{
NULL (no explicit return value).
}
\description{
This function updates the existing database file by downloading the latest version from a specified URL and saving it to the local path. It uses the \code{\link{download_db}} function internally.
}
\examples{
# Update the database file in the current working directory
update_db("initial_db.db")
}
\seealso{
\code{\link{download_db}}
}
