% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_argument.R
\name{add_argument}
\alias{add_argument}
\title{Add a filter argument to a list}
\usage{
add_argument(list, conn, variable, operator, values, statement = NULL)
}
\arguments{
\item{list}{The list to which the argument will be added.}

\item{conn}{The connection object or database connection string.}

\item{variable}{The variable name to be used in the argument.}

\item{operator}{The operator to be used in the argument (e.g., "greater", "between", "equal", "less").}

\item{values}{The values to be used in the argument.}

\item{statement}{The manual argument select statement to be used.}
}
\value{
The updated list with the added argument.
}
\description{
This function adds an argument to a list containing filter arguments later used to query databases.
The user can either specify the variable on which to filter on and the operator and value used in the filter or specify an SQL query manually.
When supplying only variable, operator and value, a SQL query will be constructed for the user and added as the next object to a list.
}
\examples{
conn <- connect_to_db(":memory:")

DBI::dbWriteTable(conn, "mtcars", mtcars)

# Initializing argument list
arguments = list()

# Using "equal" operator
arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "cyl",
 operator = "equal",
 values = c(4, 6)
)

# Using "greater" operator
arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "cyl",
 operator = "greater",
 values = 2
)

# Using "between" operator
arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "cyl",
 operator = "between",
 values = c(2, 8)
)

# Manully constructing a filter statement
manual_arguments = add_argument(
 list = arguments,
 conn = conn,
 statement = "SELECT carb FROM mtcars WHERE cyl = 4 OR cyl = 6)"
)
}
