% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_argument.R
\name{add_argument}
\alias{add_argument}
\title{Add a filter argument to a list}
\usage{
add_argument(list, conn, variable, operator, values, statement = NULL)
}
\arguments{
\item{list}{The list to which the argument will be added.}

\item{conn}{The connection object or database connection string.}

\item{variable}{The variable name to be used in the argument.}

\item{operator}{The operator to be used in the argument (e.g., "greater", "between", "equal", "less").}

\item{values}{The values to be used in the argument.}

\item{statement}{The manual argument select statement to be used.}
}
\value{
The updated list with the added argument.
}
\description{
This function adds an argument to a list based on specified conditions.
}
\examples{
conn <- connect_to_db(":memory:")

DBI::dbWriteTable(conn, "mtcars", mtcars)

arguments = list()
arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "cyl",
 operator = "equal",
 values = c(4, 6)
)

manual_arguments = add_argument(
 list = arguments,
 conn = conn,
 statement = "SELECT carb FROM mtcars WHERE cyl = 4 OR cyl = 6)"
)
}
