% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{Query Database}
\usage{
query_db(
  conn,
  arguments,
  target_vars = NULL,
  argument_relation = "and",
  target_table = "observation_table",
  full_db = TRUE
)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{arguments}{A list of filtering arguments for the query.}

\item{target_vars}{A character vector specifying the variables to be included in the query results.}

\item{argument_relation}{A character string specifying the relation between filtering arguments ("and" or "or").}

\item{target_table}{The target table in the database for querying.}

\item{full_db}{A logical value indicating whether to return the entire database. Only works when the target table is "observation_table".}
}
\value{
The query results as a data frame.
}
\description{
This function performs targeted queries on a database using specified filtering arguments and returns the query results.
}
