% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_db.R
\name{query_db}
\alias{query_db}
\title{Query Database}
\usage{
query_db(
  conn,
  arguments,
  target_vars = NULL,
  argument_relation = "and",
  target_table = "observation_table",
  full_db = TRUE
)
}
\arguments{
\item{conn}{The connection object or database connection string.}

\item{arguments}{A list of filtering arguments for the query.}

\item{target_vars}{A character vector specifying the variables to be included in the query results.}

\item{argument_relation}{A character string specifying the relation between filtering arguments ("and" or "or" or a numerical vector with the same length as the number of arguments).}

\item{target_table}{The target table in the database for querying.}

\item{full_db}{A logical value indicating whether target variables are located only in one table, or distributed among multiple tables in the database.}
}
\value{
The query results as a data frame.
}
\description{
This function performs targeted queries on a database using specified filtering arguments and returns the query results.
}
\examples{
conn <- connect_to_db(":memory:")

mtcars$mtcars_id = 1:nrow(mtcars)
PlantGrowth$plant_id = 1:nrow(PlantGrowth)
PlantGrowth$mtcars_id = 30:1

DBI::dbWriteTable(conn, "mtcars_table", mtcars)
DBI::dbWriteTable(conn, "plant_table", PlantGrowth)

arguments = list()
arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "cyl",
 operator = "equal",
 values = c(4, 6)
)

arguments = add_argument(
 list = arguments,
 conn = conn,
 variable = "weight",
 operator = "greater",
 values = 5
)

query_results = query_db(
 conn = conn,
 arguments = arguments,
 target_vars = c("mtcars_id", "plant_id", "cyl", "weight"),
 argument_relation = "and",
 target_table = "plant_table",
 full_db = TRUE
)
}
